requirements.txt
alembic==1.0.8
aniso8601==6.0.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==0.12.2
flask-heroku==0.1.9
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.5
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.14.0
marshmallow-sqlalchemy==0.13.2
psycopg2==2.7.5
psycopg2-binary==2.7.7
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
twilio==6.26.0
urllib3==1.24.1
Werkzeug==0.15.1
WTForms==2.2.1
